Remove console logs and move tests to new file
